chore(deps): bump to revm 33, latest Reth + fix zepter lint (#878)
Closes: https://github.com/tempoxyz/tempo/issues/875

Bumps Reth to latest commit:
https://github.com/paradigmxyz/reth/commit/95b8a8535b9bf0bad3222aaa6e24aa1056c6bda9,
please change if you have a policy around this!

Bumps to Revm: 33.0.0
Bumps to Alloy-Evm: 0.24.1

Note: I've ran `cargo update` to fix the lockfile

I've used `.into()` to convert `String` into `Cow<'_, str>` assuming
this will be optimized at compile time

No logic changes but please carefully review

Some internals likely have to be updated for the emitting of the events,
this has not yet been included as I don't know the details. Currently
marked as `draft` because of this. Feel free to take over the PR.

---------

Co-authored-by: Arsenii Kulikov <klkvrr@gmail.com>