fix(e2e): ensure write transaction finishes before exiting `TestingNode::stop_execution` (#1385)
When restarting a node, there might be a lingering write transaction in
the process of being committed which can result in the following edge
case (and apparently not so low chance event):

* Storage at block 27
* Reth persistence layer starts to append 28,29,30.
  * `StaticFileProvider` commits up to block 30 (successful)
* `DatabaseProviderRW` is in the process of committing up to block 30
(not finished yet)
* `TestingNode::stop_execution` is called and returns
* `TestingNode::start_execution` is eventually called again
* Reth spawns and opens a DB read-only transaction ( `DbTxRW` from above
hasnt finished yet, so this one gets a stale view)
* DB<>SF consistency check is triggered. This `DbTx` is at 27, so it
unwinds SF from 30 to 27.
  * `DbTxRW::commit` from before finishes in the background.
* Eventually on lauching the node, a new DB read-only transaction is
created, now at block 30. And it expects SF to be at 30 when it fetches
the best block -> error


Solution:
Don't return from `TestingNode::stop_execution` until a DB read-write is
available